Add None checks

+++ package.py
@@ -4,7 +4,6 @@ from functools import partial
 import json
 import os
 import sublime
-import sublime_plugin
 from threading import Thread
 
 from sublime_plugin import (
@@ -12,6 +11,7 @@ from sublime_plugin import (
     ListInputHandler,
     TextCommand,
     TextInputHandler,
+    ViewEventListener,
     WindowCommand,
 )
 
@@ -160,7 +160,7 @@ def set_layout(window):
 
         window.set_layout(layout)
 
-class TemporaryFileEventListener(sublime_plugin.ViewEventListener):
+class TemporaryFileEventListener(ViewEventListener):
     @classmethod
     def is_applicable(_, settings):
         return settings.has("tutkain_temp_file")
@@ -170,11 +170,11 @@ class TemporaryFileEventListener(sublime_plugin.ViewEventListener):
             if temp_file := self.view.settings().get("tutkain_temp_file"):
                 path = temp_file.get("path")
                 descriptor = temp_file.get("descriptor")
-                name = temp_file.get("name")
 
-                self.view.set_name(name)
+                if name := temp_file.get("name"):
+                    self.view.set_name(name)
 
-                if os.path.exists(path):
+                if path and descriptor and os.path.exists(path):
                     os.close(descriptor)
                     os.remove(path)
         except:
